+++ src/methods/index.ts
@@ -1,2 +1,0 @@
-export * from "~/methods/file.method";
-export * from "~/methods/repo.method";