+++ src/models/sync.ts
@@ -1,4 +1,4 @@
-export interface ISyncMethod {
+export interface ISyncer {
   sync(): Promise<void>;
   upload(): Promise<void>;
   download(): Promise<void>;