+++ src/services/oauth.ts
@@ -19,35 +19,34 @@ export class OAuth {
       app.get("/callback", async (req, res) => {
         try {
           const token = await this.getToken(req.param("code"), host);
+          const user = await this.getUser(token, host);
 
           res.send(`
-        <!doctype html>
-        <html lang="en">
-          <head>
-            <meta charset="utf-8">
-            <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-          </head>
-          <body>
-              <h1>Success! You may now close this tab.</h1>
-              <style>
-                html, body {
-                  background-color: #1a1a1a;
-                  color: #c3c3c3;
-                  display: flex;
-                  justify-content: center;
-                  align-items: center;
-                  height: 100%;
-                  width: 100%;
-                  margin: 0;
-                }
-              </style>
-          </body>
-        </html>
-        `);
+          <!doctype html>
+          <html lang="en">
+            <head>
+              <meta charset="utf-8">
+              <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+            </head>
+            <body>
+                <h1>Success! You may now close this tab.</h1>
+                <style>
+                  html, body {
+                    background-color: #1a1a1a;
+                    color: #c3c3c3;
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    height: 100%;
+                    width: 100%;
+                    margin: 0;
+                  }
+                </style>
+            </body>
+          </html>
+          `);
 
           server.close();
-
-          const user = await this.getUser(token, host);
 
           this.saveCredentials(token, user);