+++ src/syncers/index.ts
@@ -1,0 +1,2 @@
+export * from "~/syncers/file";
+export * from "~/syncers/repo";