+++ src/test/services/custom-file.test.ts
@@ -4,13 +4,10 @@ import { resolve } from "path";
 import { Uri } from "vscode";
 import { CustomFiles, Environment, FS } from "~/services";
 
-jest.mock("~/services/localization.service.ts");
-jest.mock("~/models/state.model.ts");
+jest.mock("~/services/localization.ts");
+jest.mock("~/models/state.ts");
 
-const cleanupPath = resolve(
-  tmpdir(),
-  "syncify-jest/utility/custom-file.service"
-);
+const cleanupPath = resolve(tmpdir(), "syncify-jest/services/custom-file");
 const pathToSource = `${cleanupPath}/source`;
 const pathToRegistered = `${cleanupPath}/registered`;