+++ src/test/services/fs.test.ts
@@ -3,10 +3,10 @@ import { tmpdir } from "os";
 import { resolve } from "path";
 import { FS, Settings } from "~/services";
 
-jest.mock("~/services/localization.service.ts");
-jest.mock("~/models/state.model.ts");
+jest.mock("~/services/localization.ts");
+jest.mock("~/models/state.ts");
 
-const cleanupPath = resolve(tmpdir(), "syncify-jest/services/fs.service");
+const cleanupPath = resolve(tmpdir(), "syncify-jest/services/fs");
 const testFolder = `${cleanupPath}/test`;
 
 beforeEach(() => Promise.all([ensureDir(testFolder)]));